pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "FastAPI App"
authors = ["Keiken <anubhav@threye.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-decouple = "^3.8"
pydantic-settings = "^2.2.1"
beanie = "^1.25.0"
email-validator = "^2.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
